+++ library/core/common/src/main/kotlin/com/mihaicristiancondrea/android/libs/apptoolkit/core/common/utils/constants/colorscheme/PaletteConstants.kt
@@ -34,6 +34,7 @@ object DynamicPaletteVariant {
 object StaticPaletteIds {
     const val MONOCHROME = "monochrome"
     const val GOOGLE_BLUE = "blue"
+    const val PURPLE = "purple"
     const val ANDROID = "android"
     const val GREEN = "green"
     const val RED = "red"
@@ -49,6 +50,7 @@ object StaticPaletteIds {
     private val supportedOrder = listOf(
         MONOCHROME,
         GOOGLE_BLUE,
+        PURPLE,
         ANDROID,
         GREEN,
         RED,